Refactor repo and namespace hooks
